# backend/scripts/swagger_v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DataLabellingSupportSystem.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Roles.CreateRoleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Roles.CreateRoleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Roles.CreateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/roles/{roleId}": {
      "put": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Roles.UpdateRoleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Roles.UpdateRoleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Roles.UpdateRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Users.CreateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Users.CreateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Users.CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Users.UpdateUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Users.UpdateUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Users.UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse"
            },
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Collections.Generic.List_DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse"
            },
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Requests.Roles.CreateRoleRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Roles.UpdateRoleRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Users.CreateUserRequest": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "identifyNumber": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Users.UpdateUserRequest": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "identifyNumber": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Responses.Roles.RoleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Responses.Users.UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "identifyNumber": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste only the JWT access token here. Swagger will send: Authorization: Bearer <token>",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}